Move common functions to common.py
This prepares us for splitting up the tests in test_flexvolume.py.

Moved functions:
- create_and_wait_pod
- delete_pod
- read_volume_data
- write_volume_data
- size_to_string
